requirements.txt
django
django-nested-admin
django-rest-framework
django-jazzmin